+++ spec/rubocop/cop/rspec/align_left_let_brace_spec.rb
@@ -43,13 +43,13 @@ RSpec.describe RuboCop::Cop::RSpec::AlignLeftLetBrace do
   end
   # rubocop:enable RSpec/ExampleLength
 
-  it 'does not register offense for let with proc argument' do
+  it 'does not register an offense for let with proc argument' do
     expect_no_offenses(<<-RUBY)
       let(:user, &args[:build_user])
     RUBY
   end
 
-  it 'works with empty file' do
+  it 'does not register an offense with empty file' do
     expect_no_offenses('')
   end
 end